+++ src/main/java/com/pichincha/java_tcs_jlcastro/service/MovementsService.java
@@ -8,7 +8,7 @@ import com.pichincha.java_tcs_jlcastro.exception.NoResourceException;
 public interface MovementsService {
     public List<MovementsDto> getAllMovementss();
     public MovementsDto createMovements(MovementsDto MovementsDto);
-    public MovementsDto updateMovements(Long id, MovementsDto MovementsDto) throws NoResourceException;
-    public void deleteMovementse(Long id) throws NoResourceException;
+    public MovementsDto updateMovements(Integer id, MovementsDto MovementsDto) throws NoResourceException;
+    public void deleteMovementse(Integer id) throws NoResourceException;
 
 }
